requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
jira>=3.5.1
atlassian-python-api>=3.30.0
PyGithub>=1.55
pydantic>=1.8.0 